Move choice sets from utilities.choices to netbox.choices
