+++ docs/plugins/development/navigation.md
@@ -49,8 +49,8 @@ menu_items = (item1, item2, item3)
 Each menu item represents a link and (optionally) a set of buttons comprising one entry in NetBox's navigation menu. Menu items are defined as PluginMenuItem instances. An example is shown below.
 
 ```python title="navigation.py"
+from netbox.choices import ButtonColorChoices
 from netbox.plugins import PluginMenuButton, PluginMenuItem
-from utilities.choices import ButtonColorChoices
 
 item1 = PluginMenuItem(
     link='plugins:myplugin:myview',
