+++ netbox/netbox/navigation/menu.py
@@ -1,7 +1,6 @@
 from django.utils.translation import gettext_lazy as _
 
 from netbox.registry import registry
-from utilities.choices import ButtonColorChoices
 from . import *
 
 #
